+++ cmd/porter/installations.go
@@ -3,6 +3,7 @@ package main
 import (
 	"get.porter.sh/porter/pkg/porter"
 	"github.com/spf13/cobra"
+	"github.com/spf13/pflag"
 )
 
 func buildInstallationCommands(p *porter.Porter) *cobra.Command {
@@ -23,6 +24,10 @@ func buildInstallationCommands(p *porter.Porter) *cobra.Command {
 	cmd.AddCommand(buildInstallationDeleteCommand(p))
 	cmd.AddCommand(buildInstallationLogCommands(p))
 	cmd.AddCommand(buildInstallationRunsCommands(p))
+	cmd.AddCommand(buildInstallationInstallCommand(p))
+	cmd.AddCommand(buildInstallationUpgradeCommand(p))
+	cmd.AddCommand(buildInstallationInvokeCommand(p))
+	cmd.AddCommand(buildInstallationUninstallCommand(p))
 
 	return cmd
 }
@@ -205,4 +210,257 @@ func buildInstallationRunsListCommand(p *porter.Porter) *cobra.Command {
 		"Specify an output format.  Allowed values: plaintext, json, yaml")
 
 	return &cmd
+}
+
+func buildInstallationInstallCommand(p *porter.Porter) *cobra.Command {
+	opts := porter.NewInstallOptions()
+	cmd := &cobra.Command{
+		Use:   "install [INSTALLATION]",
+		Short: "Create a new installation of a bundle",
+		Long: `Create a new installation of a bundle.
+
+The first argument is the name of the installation to create. This defaults to the name of the bundle. 
+
+Once a bundle has been successfully installed, the install action cannot be repeated. This is a precaution to avoid accidentally overwriting an existing installation. If you need to re-run install, which is common when authoring a bundle, you can use the --force flag to by-pass this check.
+
+Porter uses the docker driver as the default runtime for executing a bundle's invocation image, but an alternate driver may be supplied via '--driver/-d' or the PORTER_RUNTIME_DRIVER environment variable.
+For example, the 'debug' driver may be specified, which simply logs the info given to it and then exits.
+
+The docker driver runs the bundle container using the local Docker host. To use a remote Docker host, set the following environment variables:
+  DOCKER_HOST (required)
+  DOCKER_TLS_VERIFY (optional)
+  DOCKER_CERT_PATH (optional)
+`,
+		Example: `  porter bundle install
+  porter bundle install MyAppFromReference --reference ghcr.io/getporter/examples/kubernetes:v0.2.0 --namespace dev
+  porter bundle install --reference localhost:5000/ghcr.io/getporter/examples/kubernetes:v0.2.0 --insecure-registry --force
+  porter bundle install MyAppInDev --file myapp/bundle.json
+  porter bundle install --parameter-set azure --param test-mode=true --param header-color=blue
+  porter bundle install --cred azure --cred kubernetes
+  porter bundle install --driver debug
+  porter bundle install --label env=dev --label owner=myuser
+`,
+		PreRunE: func(cmd *cobra.Command, args []string) error {
+			return opts.Validate(cmd.Context(), args, p)
+		},
+		RunE: func(cmd *cobra.Command, args []string) error {
+			return p.InstallBundle(cmd.Context(), opts)
+		},
+	}
+
+	f := cmd.Flags()
+	f.StringVarP(&opts.File, "file", "f", "",
+		"Path to the porter manifest file. Defaults to the bundle in the current directory.")
+	f.StringVar(&opts.CNABFile, "cnab-file", "",
+		"Path to the CNAB bundle.json file.")
+	f.StringArrayVarP(&opts.ParameterSets, "parameter-set", "p", nil,
+		"Name of a parameter set for the bundle. It should be a named set of parameters and may be specified multiple times.")
+	f.StringArrayVar(&opts.Params, "param", nil,
+		"Define an individual parameter in the form NAME=VALUE. Overrides parameters otherwise set via --parameter-set. May be specified multiple times.")
+	f.StringArrayVarP(&opts.CredentialIdentifiers, "cred", "c", nil,
+		"Credential to use when installing the bundle. It should be a named set of credentials and may be specified multiple times.")
+	f.StringVarP(&opts.Driver, "driver", "d", porter.DefaultDriver,
+		"Specify a driver to use. Allowed values: docker, debug")
+	f.StringVarP(&opts.Namespace, "namespace", "n", "",
+		"Create the installation in the specified namespace. Defaults to the global namespace.")
+	f.StringSliceVarP(&opts.Labels, "label", "l", nil,
+		"Associate the specified labels with the installation. May be specified multiple times.")
+	addBundleActionFlags(f, opts)
+
+	// Allow configuring the --driver flag with runtime-driver, to avoid conflicts with other commands
+	cmd.Flag("driver").Annotations = map[string][]string{
+		"viper-key": {"runtime-driver"},
+	}
+	return cmd
+}
+
+func buildInstallationUpgradeCommand(p *porter.Porter) *cobra.Command {
+	opts := porter.NewUpgradeOptions()
+	cmd := &cobra.Command{
+		Use:   "upgrade [INSTALLATION]",
+		Short: "Upgrade an installation",
+		Long: `Upgrade an installation.
+
+The first argument is the installation name to upgrade. This defaults to the name of the bundle.
+
+Porter uses the docker driver as the default runtime for executing a bundle's invocation image, but an alternate driver may be supplied via '--driver/-d' or the PORTER_RUNTIME_DRIVER environment variable.
+For example, the 'debug' driver may be specified, which simply logs the info given to it and then exits.
+
+The docker driver runs the bundle container using the local Docker host. To use a remote Docker host, set the following environment variables:
+  DOCKER_HOST (required)
+  DOCKER_TLS_VERIFY (optional)
+  DOCKER_CERT_PATH (optional)
+`,
+		Example: `  porter bundle upgrade --version 0.2.0
+  porter bundle upgrade --reference ghcr.io/getporter/examples/kubernetes:v0.2.0
+  porter bundle upgrade --reference localhost:5000/ghcr.io/getporter/examples/kubernetes:v0.2.0 --insecure-registry --force
+  porter bundle upgrade MyAppInDev --file myapp/bundle.json
+  porter bundle upgrade --parameter-set azure --param test-mode=true --param header-color=blue
+  porter bundle upgrade --cred azure --cred kubernetes
+  porter bundle upgrade --driver debug
+`,
+		PreRunE: func(cmd *cobra.Command, args []string) error {
+			return opts.Validate(cmd.Context(), args, p)
+		},
+		RunE: func(cmd *cobra.Command, args []string) error {
+			return p.UpgradeBundle(cmd.Context(), opts)
+		},
+	}
+
+	f := cmd.Flags()
+	f.StringVarP(&opts.File, "file", "f", "",
+		"Path to the porter manifest file. Defaults to the bundle in the current directory.")
+	f.StringVar(&opts.CNABFile, "cnab-file", "",
+		"Path to the CNAB bundle.json file.")
+	f.StringArrayVarP(&opts.ParameterSets, "parameter-set", "p", nil,
+		"Name of a parameter set file for the bundle. May be either a named set of parameters or a filepath, and specified multiple times.")
+	f.StringArrayVar(&opts.Params, "param", nil,
+		"Define an individual parameter in the form NAME=VALUE. Overrides parameters otherwise set via --parameter-set. May be specified multiple times.")
+	f.StringArrayVarP(&opts.CredentialIdentifiers, "cred", "c", nil,
+		"Credential to use when installing the bundle. It should be a named set of credentials and may be specified multiple times.")
+	f.StringVarP(&opts.Driver, "driver", "d", porter.DefaultDriver,
+		"Specify a driver to use. Allowed values: docker, debug")
+	f.StringVarP(&opts.Namespace, "namespace", "n", "",
+		"Namespace of the specified installation. Defaults to the global namespace.")
+	f.StringVar(&opts.Version, "version", "",
+		"Version to which the installation should be upgraded. This represents the version of the bundle, which assumes the convention of setting the bundle tag to its version.")
+	addBundleActionFlags(f, opts)
+
+	// Allow configuring the --driver flag with runtime-driver, to avoid conflicts with other commands
+	cmd.Flag("driver").Annotations = map[string][]string{
+		"viper-key": {"runtime-driver"},
+	}
+	return cmd
+}
+
+func buildInstallationInvokeCommand(p *porter.Porter) *cobra.Command {
+	opts := porter.NewInvokeOptions()
+	cmd := &cobra.Command{
+		Use:   "invoke [INSTALLATION] --action ACTION",
+		Short: "Invoke a custom action on an installation",
+		Long: `Invoke a custom action on an installation.
+
+The first argument is the installation name upon which to invoke the action. This defaults to the name of the bundle.
+
+Porter uses the docker driver as the default runtime for executing a bundle's invocation image, but an alternate driver may be supplied via '--driver/-d' or the PORTER_RUNTIME_DRIVER environment variable.
+For example, the 'debug' driver may be specified, which simply logs the info given to it and then exits.
+
+The docker driver runs the bundle container using the local Docker host. To use a remote Docker host, set the following environment variables:
+  DOCKER_HOST (required)
+  DOCKER_TLS_VERIFY (optional)
+  DOCKER_CERT_PATH (optional)
+`,
+		Example: `  porter bundle invoke --action ACTION
+  porter bundle invoke --reference ghcr.io/getporter/examples/kubernetes:v0.2.0
+  porter bundle invoke --reference localhost:5000/ghcr.io/getporter/examples/kubernetes:v0.2.0 --insecure-registry --force
+  porter bundle invoke --action ACTION MyAppInDev --file myapp/bundle.json
+  porter bundle invoke --action ACTION  --parameter-set azure --param test-mode=true --param header-color=blue
+  porter bundle invoke --action ACTION --cred azure --cred kubernetes
+  porter bundle invoke --action ACTION --driver debug
+`,
+		PreRunE: func(cmd *cobra.Command, args []string) error {
+			return opts.Validate(cmd.Context(), args, p)
+		},
+		RunE: func(cmd *cobra.Command, args []string) error {
+			return p.InvokeBundle(cmd.Context(), opts)
+		},
+	}
+
+	f := cmd.Flags()
+	f.StringVar(&opts.Action, "action", "",
+		"Custom action name to invoke.")
+	f.StringVarP(&opts.File, "file", "f", "",
+		"Path to the porter manifest file. Defaults to the bundle in the current directory.")
+	f.StringVar(&opts.CNABFile, "cnab-file", "",
+		"Path to the CNAB bundle.json file.")
+	f.StringArrayVarP(&opts.ParameterSets, "parameter-set", "p", nil,
+		"Name of a parameter set file for the bundle. May be either a named set of parameters or a filepath, and specified multiple times.")
+	f.StringArrayVar(&opts.Params, "param", nil,
+		"Define an individual parameter in the form NAME=VALUE. Overrides parameters otherwise set via --parameter-set. May be specified multiple times.")
+	f.StringArrayVarP(&opts.CredentialIdentifiers, "cred", "c", nil,
+		"Credential to use when installing the bundle. May be either a named set of credentials or a filepath, and specified multiple times.")
+	f.StringVarP(&opts.Driver, "driver", "d", porter.DefaultDriver,
+		"Specify a driver to use. Allowed values: docker, debug")
+	f.StringVarP(&opts.Namespace, "namespace", "n", "",
+		"Namespace of the specified installation. Defaults to the global namespace.")
+	addBundleActionFlags(f, opts)
+
+	// Allow configuring the --driver flag with runtime-driver, to avoid conflicts with other commands
+	cmd.Flag("driver").Annotations = map[string][]string{
+		"viper-key": {"runtime-driver"},
+	}
+	return cmd
+}
+
+func buildInstallationUninstallCommand(p *porter.Porter) *cobra.Command {
+	opts := porter.NewUninstallOptions()
+	cmd := &cobra.Command{
+		Use:   "uninstall [INSTALLATION]",
+		Short: "Uninstall an installation",
+		Long: `Uninstall an installation
+
+The first argument is the installation name to uninstall. This defaults to the name of the bundle.
+
+Porter uses the docker driver as the default runtime for executing a bundle's invocation image, but an alternate driver may be supplied via '--driver/-d'' or the PORTER_RUNTIME_DRIVER environment variable.
+For example, the 'debug' driver may be specified, which simply logs the info given to it and then exits.
+
+The docker driver runs the bundle container using the local Docker host. To use a remote Docker host, set the following environment variables:
+  DOCKER_HOST (required)
+  DOCKER_TLS_VERIFY (optional)
+  DOCKER_CERT_PATH (optional)
+`,
+		Example: `  porter bundle uninstall
+  porter bundle uninstall --reference ghcr.io/getporter/examples/kubernetes:v0.2.0
+  porter bundle uninstall --reference localhost:5000/ghcr.io/getporter/examples/kubernetes:v0.2.0 --insecure-registry --force
+  porter bundle uninstall MyAppInDev --file myapp/bundle.json
+  porter bundle uninstall --parameter-set azure --param test-mode=true --param header-color=blue
+  porter bundle uninstall --cred azure --cred kubernetes
+  porter bundle uninstall --driver debug
+  porter bundle uninstall --delete
+  porter bundle uninstall --force-delete
+`,
+		PreRunE: func(cmd *cobra.Command, args []string) error {
+			return opts.Validate(cmd.Context(), args, p)
+		},
+		RunE: func(cmd *cobra.Command, args []string) error {
+			return p.UninstallBundle(cmd.Context(), opts)
+		},
+	}
+
+	f := cmd.Flags()
+	f.StringVarP(&opts.File, "file", "f", "",
+		"Path to the porter manifest file. Defaults to the bundle in the current directory. Optional unless a newer version of the bundle should be used to uninstall the bundle.")
+	f.StringVar(&opts.CNABFile, "cnab-file", "",
+		"Path to the CNAB bundle.json file.")
+	f.StringArrayVarP(&opts.ParameterSets, "parameter-set", "p", nil,
+		"Name of a parameter set file for the bundle. May be either a named set of parameters or a filepath, and specified multiple times.")
+	f.StringArrayVar(&opts.Params, "param", nil,
+		"Define an individual parameter in the form NAME=VALUE. Overrides parameters otherwise set via --parameter-set. May be specified multiple times.")
+	f.StringArrayVarP(&opts.CredentialIdentifiers, "cred", "c", nil,
+		"Credential to use when uninstalling the bundle. May be either a named set of credentials or a filepath, and specified multiple times.")
+	f.StringVarP(&opts.Driver, "driver", "d", porter.DefaultDriver,
+		"Specify a driver to use. Allowed values: docker, debug")
+	f.BoolVar(&opts.Delete, "delete", false,
+		"Delete all records associated with the installation, assuming the uninstall action succeeds")
+	f.BoolVar(&opts.ForceDelete, "force-delete", false,
+		"UNSAFE. Delete all records associated with the installation, even if uninstall fails. This is intended for cleaning up test data and is not recommended for production environments.")
+	f.StringVarP(&opts.Namespace, "namespace", "n", "",
+		"Namespace of the specified installation. Defaults to the global namespace.")
+	addBundleActionFlags(f, opts)
+
+	// Allow configuring the --driver flag with runtime-driver, to avoid conflicts with other commands
+	cmd.Flag("driver").Annotations = map[string][]string{
+		"viper-key": {"runtime-driver"},
+	}
+	return cmd
+}
+
+// Add flags for command that execute a bundle (install, upgrade, invoke and uninstall)
+func addBundleActionFlags(f *pflag.FlagSet, actionOpts porter.BundleAction) {
+	opts := actionOpts.GetOptions()
+	addBundlePullFlags(f, &opts.BundlePullOptions)
+	f.BoolVar(&opts.AllowDockerHostAccess, "allow-docker-host-access", false,
+		"Controls if the bundle should have access to the host's Docker daemon with elevated privileges. See https://getporter.org/configuration/#allow-docker-host-access for the full implications of this flag.")
+	f.BoolVar(&opts.NoLogs, "no-logs", false,
+		"Do not persist the bundle execution logs")
 }
