+++ cmd/porter/main_test.go
@@ -23,8 +23,8 @@ func TestCommandWiring(t *testing.T) {
 		"bundles",
 		"bundle create",
 		"bundle build",
-		"bundle install",
-		"bundle uninstall",
+		"installation install",
+		"installation uninstall",
 		"mixins",
 		"mixins list",
 		"plugins list",
