--- docs/content/cli/_index.md
@@ -1,4 +1,0 @@
----
-Title: Porter Commands
-Description: The porter -help indexed for your SEO enjoyment
----
