+++ docs/content/cli/bundles.md
@@ -39,9 +39,5 @@ Try our QuickStart https://getporter.org/quickstart to learn how to use Porter.
 * [porter bundles create](/cli/porter_bundles_create/)	 - Create a bundle
 * [porter bundles explain](/cli/porter_bundles_explain/)	 - Explain a bundle
 * [porter bundles inspect](/cli/porter_bundles_inspect/)	 - Inspect a bundle
-* [porter bundles install](/cli/porter_bundles_install/)	 - Create a new installation of a bundle
-* [porter bundles invoke](/cli/porter_bundles_invoke/)	 - Invoke a custom action on an installation
 * [porter bundles lint](/cli/porter_bundles_lint/)	 - Lint a bundle
-* [porter bundles uninstall](/cli/porter_bundles_uninstall/)	 - Uninstall an installation
-* [porter bundles upgrade](/cli/porter_bundles_upgrade/)	 - Upgrade an installation
