+++ docs/content/cli/install.md
@@ -31,14 +31,14 @@ porter install [INSTALLATION] [flags]
 ### Examples
 
 ```
-  porter install
-  porter install MyAppFromReference --reference ghcr.io/getporter/examples/kubernetes:v0.2.0 --namespace dev
-  porter install --reference localhost:5000/ghcr.io/getporter/examples/kubernetes:v0.2.0 --insecure-registry --force
-  porter install MyAppInDev --file myapp/bundle.json
-  porter install --parameter-set azure --param test-mode=true --param header-color=blue
-  porter install --cred azure --cred kubernetes
-  porter install --driver debug
-  porter install --label env=dev --label owner=myuser
+  porter bundle install
+  porter bundle install MyAppFromReference --reference ghcr.io/getporter/examples/kubernetes:v0.2.0 --namespace dev
+  porter bundle install --reference localhost:5000/ghcr.io/getporter/examples/kubernetes:v0.2.0 --insecure-registry --force
+  porter bundle install MyAppInDev --file myapp/bundle.json
+  porter bundle install --parameter-set azure --param test-mode=true --param header-color=blue
+  porter bundle install --cred azure --cred kubernetes
+  porter bundle install --driver debug
+  porter bundle install --label env=dev --label owner=myuser
 
 ```
