+++ docs/content/cli/installations_invoke.md
@@ -1,19 +1,17 @@
 ---
-title: "porter bundles install"
-slug: porter_bundles_install
-url: /cli/porter_bundles_install/
+title: "porter installations invoke"
+slug: porter_installations_invoke
+url: /cli/porter_installations_invoke/
 ---
-## porter bundles install
+## porter installations invoke
 
-Create a new installation of a bundle
+Invoke a custom action on an installation
 
 ### Synopsis
 
-Create a new installation of a bundle.
+Invoke a custom action on an installation.
 
-The first argument is the name of the installation to create. This defaults to the name of the bundle. 
-
-Once a bundle has been successfully installed, the install action cannot be repeated. This is a precaution to avoid accidentally overwriting an existing installation. If you need to re-run install, which is common when authoring a bundle, you can use the --force flag to by-pass this check.
+The first argument is the installation name upon which to invoke the action. This defaults to the name of the bundle.
 
 Porter uses the docker driver as the default runtime for executing a bundle's invocation image, but an alternate driver may be supplied via '--driver/-d' or the PORTER_RUNTIME_DRIVER environment variable.
 For example, the 'debug' driver may be specified, which simply logs the info given to it and then exits.
@@ -25,39 +23,38 @@ The docker driver runs the bundle container using the local Docker host. To use 
 
 
 ```
-porter bundles install [INSTALLATION] [flags]
+porter installations invoke [INSTALLATION] --action ACTION [flags]
 ```
 
 ### Examples
 
 ```
-  porter bundle install
-  porter bundle install MyAppFromReference --reference ghcr.io/getporter/examples/kubernetes:v0.2.0 --namespace dev
-  porter bundle install --reference localhost:5000/ghcr.io/getporter/examples/kubernetes:v0.2.0 --insecure-registry --force
-  porter bundle install MyAppInDev --file myapp/bundle.json
-  porter bundle install --parameter-set azure --param test-mode=true --param header-color=blue
-  porter bundle install --cred azure --cred kubernetes
-  porter bundle install --driver debug
-  porter bundle install --label env=dev --label owner=myuser
+  porter bundle invoke --action ACTION
+  porter bundle invoke --reference ghcr.io/getporter/examples/kubernetes:v0.2.0
+  porter bundle invoke --reference localhost:5000/ghcr.io/getporter/examples/kubernetes:v0.2.0 --insecure-registry --force
+  porter bundle invoke --action ACTION MyAppInDev --file myapp/bundle.json
+  porter bundle invoke --action ACTION  --parameter-set azure --param test-mode=true --param header-color=blue
+  porter bundle invoke --action ACTION --cred azure --cred kubernetes
+  porter bundle invoke --action ACTION --driver debug
 
 ```
 
 ### Options
 
 ```
+      --action string               Custom action name to invoke.
       --allow-docker-host-access    Controls if the bundle should have access to the host's Docker daemon with elevated privileges. See https://getporter.org/configuration/#allow-docker-host-access for the full implications of this flag.
       --cnab-file string            Path to the CNAB bundle.json file.
-  -c, --cred stringArray            Credential to use when installing the bundle. It should be a named set of credentials and may be specified multiple times.
+  -c, --cred stringArray            Credential to use when installing the bundle. May be either a named set of credentials or a filepath, and specified multiple times.
   -d, --driver string               Specify a driver to use. Allowed values: docker, debug (default "docker")
   -f, --file string                 Path to the porter manifest file. Defaults to the bundle in the current directory.
       --force                       Force a fresh pull of the bundle
-  -h, --help                        help for install
+  -h, --help                        help for invoke
       --insecure-registry           Don't require TLS for the registry
-  -l, --label strings               Associate the specified labels with the installation. May be specified multiple times.
-  -n, --namespace string            Create the installation in the specified namespace. Defaults to the global namespace.
+  -n, --namespace string            Namespace of the specified installation. Defaults to the global namespace.
       --no-logs                     Do not persist the bundle execution logs
       --param stringArray           Define an individual parameter in the form NAME=VALUE. Overrides parameters otherwise set via --parameter-set. May be specified multiple times.
-  -p, --parameter-set stringArray   Name of a parameter set for the bundle. It should be a named set of parameters and may be specified multiple times.
+  -p, --parameter-set stringArray   Name of a parameter set file for the bundle. May be either a named set of parameters or a filepath, and specified multiple times.
   -r, --reference string            Use a bundle in an OCI registry specified by the given reference.
 ```
 
@@ -71,5 +68,5 @@ porter bundles install [INSTALLATION] [flags]
 
 ### SEE ALSO
 
-* [porter bundles](/cli/porter_bundles/)	 - Bundle commands
+* [porter installations](/cli/porter_installations/)	 - Installation commands
