+++ docs/content/cli/installations_upgrade.md
@@ -1,17 +1,17 @@
 ---
-title: "porter bundles invoke"
-slug: porter_bundles_invoke
-url: /cli/porter_bundles_invoke/
+title: "porter installations upgrade"
+slug: porter_installations_upgrade
+url: /cli/porter_installations_upgrade/
 ---
-## porter bundles invoke
+## porter installations upgrade
 
-Invoke a custom action on an installation
+Upgrade an installation
 
 ### Synopsis
 
-Invoke a custom action on an installation.
+Upgrade an installation.
 
-The first argument is the installation name upon which to invoke the action. This defaults to the name of the bundle.
+The first argument is the installation name to upgrade. This defaults to the name of the bundle.
 
 Porter uses the docker driver as the default runtime for executing a bundle's invocation image, but an alternate driver may be supplied via '--driver/-d' or the PORTER_RUNTIME_DRIVER environment variable.
 For example, the 'debug' driver may be specified, which simply logs the info given to it and then exits.
@@ -23,39 +23,39 @@ The docker driver runs the bundle container using the local Docker host. To use 
 
 
 ```
-porter bundles invoke [INSTALLATION] --action ACTION [flags]
+porter installations upgrade [INSTALLATION] [flags]
 ```
 
 ### Examples
 
 ```
-  porter bundle invoke --action ACTION
-  porter bundle invoke --reference ghcr.io/getporter/examples/kubernetes:v0.2.0
-  porter bundle invoke --reference localhost:5000/ghcr.io/getporter/examples/kubernetes:v0.2.0 --insecure-registry --force
-  porter bundle invoke --action ACTION MyAppInDev --file myapp/bundle.json
-  porter bundle invoke --action ACTION  --parameter-set azure --param test-mode=true --param header-color=blue
-  porter bundle invoke --action ACTION --cred azure --cred kubernetes
-  porter bundle invoke --action ACTION --driver debug
+  porter bundle upgrade --version 0.2.0
+  porter bundle upgrade --reference ghcr.io/getporter/examples/kubernetes:v0.2.0
+  porter bundle upgrade --reference localhost:5000/ghcr.io/getporter/examples/kubernetes:v0.2.0 --insecure-registry --force
+  porter bundle upgrade MyAppInDev --file myapp/bundle.json
+  porter bundle upgrade --parameter-set azure --param test-mode=true --param header-color=blue
+  porter bundle upgrade --cred azure --cred kubernetes
+  porter bundle upgrade --driver debug
 
 ```
 
 ### Options
 
 ```
-      --action string               Custom action name to invoke.
       --allow-docker-host-access    Controls if the bundle should have access to the host's Docker daemon with elevated privileges. See https://getporter.org/configuration/#allow-docker-host-access for the full implications of this flag.
       --cnab-file string            Path to the CNAB bundle.json file.
-  -c, --cred stringArray            Credential to use when installing the bundle. May be either a named set of credentials or a filepath, and specified multiple times.
+  -c, --cred stringArray            Credential to use when installing the bundle. It should be a named set of credentials and may be specified multiple times.
   -d, --driver string               Specify a driver to use. Allowed values: docker, debug (default "docker")
   -f, --file string                 Path to the porter manifest file. Defaults to the bundle in the current directory.
       --force                       Force a fresh pull of the bundle
-  -h, --help                        help for invoke
+  -h, --help                        help for upgrade
       --insecure-registry           Don't require TLS for the registry
   -n, --namespace string            Namespace of the specified installation. Defaults to the global namespace.
       --no-logs                     Do not persist the bundle execution logs
       --param stringArray           Define an individual parameter in the form NAME=VALUE. Overrides parameters otherwise set via --parameter-set. May be specified multiple times.
   -p, --parameter-set stringArray   Name of a parameter set file for the bundle. May be either a named set of parameters or a filepath, and specified multiple times.
   -r, --reference string            Use a bundle in an OCI registry specified by the given reference.
+      --version string              Version to which the installation should be upgraded. This represents the version of the bundle, which assumes the convention of setting the bundle tag to its version.
 ```
 
 ### Options inherited from parent commands
@@ -68,5 +68,5 @@ porter bundles invoke [INSTALLATION] --action ACTION [flags]
 
 ### SEE ALSO
 
-* [porter bundles](/cli/porter_bundles/)	 - Bundle commands
+* [porter installations](/cli/porter_installations/)	 - Installation commands
