+++ docs/content/cli/invoke.md
@@ -29,13 +29,13 @@ porter invoke [INSTALLATION] --action ACTION [flags]
 ### Examples
 
 ```
-  porter invoke --action ACTION
-  porter invoke --reference ghcr.io/getporter/examples/kubernetes:v0.2.0
-  porter invoke --reference localhost:5000/ghcr.io/getporter/examples/kubernetes:v0.2.0 --insecure-registry --force
-  porter invoke --action ACTION MyAppInDev --file myapp/bundle.json
-  porter invoke --action ACTION  --parameter-set azure --param test-mode=true --param header-color=blue
-  porter invoke --action ACTION --cred azure --cred kubernetes
-  porter invoke --action ACTION --driver debug
+  porter bundle invoke --action ACTION
+  porter bundle invoke --reference ghcr.io/getporter/examples/kubernetes:v0.2.0
+  porter bundle invoke --reference localhost:5000/ghcr.io/getporter/examples/kubernetes:v0.2.0 --insecure-registry --force
+  porter bundle invoke --action ACTION MyAppInDev --file myapp/bundle.json
+  porter bundle invoke --action ACTION  --parameter-set azure --param test-mode=true --param header-color=blue
+  porter bundle invoke --action ACTION --cred azure --cred kubernetes
+  porter bundle invoke --action ACTION --driver debug
 
 ```
